Use DeviancePlugin not DeviancePluginS in icSet() and dicSetS()
See Andrew email 2020-10-27 15:23

When I did the HMC stuff it was necessary to map parameters to the whole real line. I found I could use the same mapping for DIC. How is the new way related to the old? For proportions I now do the logistic transformation so it should give the same answer as a logistic model with stochastic parents. For the mean of the  poison I do log transformation so it should be the same as stochastic patents in a log linear model. Where I get lost is with say precision matrices or vectors of proportions. Mapping them to the whole real line is tougher...